requirements.txt
atproto==0.0.58
requests==2.32.3
pycryptodome==3.23.0